requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
transformers==4.36.2
ffmpeg-python==0.2.0
